third_party/blink/web_tests/fast/dom/Window/window-postmessage-clone-deep-array-expected.txt

Tests that we support cloning deep(ish) arrays.
FAIL: 'postMessage(deepArray)' should not throw but threw RangeError: Failed to execute 'postMessage' on 'Window': Maximum call stack size exceeded
PASS: eventData is done of type string
